native_service: fixes for freebasic rakehelp during clobber of projects.
git-svn-id: svn://rubyforge.org/var/svn/mongrel/branches/win32-native_service@375 19e92222-5c0b-0410-8929-a290d50e31e9

+++ projects/mongrel_service/tools/freebasic.rb
@@ -231,10 +231,13 @@ module FreeBASIC
           # now the sources files
           # avoid attempt to remove the file two times (this is a bug in Rake)
           @sources.each do |src|
-            target = compiled_form(src)
-            unless CLOBBER.include?(target)
-              CLOBBER.include(target)
-              rm target rescue nil
+            # exclude compiled source files (c obj).
+            unless src =~ /o$/
+              target = compiled_form(src)
+              unless CLOBBER.include?(target)
+                CLOBBER.include(target)
+                rm target rescue nil
+              end
             end
           end
         end
